pyproject.toml
[project]
name = "xproject"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "clip",
    "coremltools>=9.0",
    "easyocr>=1.7.2",
    "kmbox-net-rust",
    "mss>=10.1.0",
    "onnx>=1.20.1",
    "onnxruntime>=1.24.1",
    "onnxslim>=0.1.85",
    "openai-clip>=1.0.1",
    "opencv-python>=4.13.0.92",
    "pynput>=1.8.1",
    "rich>=14.3.3",
    "setuptools>=82.0.0",
    "transitions>=0.9.3",
    "ultralytics>=8.4.12",
]

[tool.uv.sources]
clip = { git = "https://github.com/ultralytics/CLIP.git" }
